requirements.txt
aiogram==3.0.0b7
Django==4.1.7
django-phonenumber-field==7.0.2
django-redis==5.2.0
ipython==8.11.0
python-dotenv==1.0.0
phonenumbers==8.13.7
uvicorn==0.20.0
